Seperate repositories is silly
This repo was originally concieved as a personal overlay to the Funtoo
portage tree. However, Funtoo's portage tree fairly closely tracks the
Gentoo portage tree, with only a handful of ebuilds dfferent from
Gentoo. Also, it is generally a good idea to write the ebuilds so they
work equally on both Gentoo and Funtoo -- portabilty and all that good
stuff.

Thus, given the above, it makes no sense to keep seperate repos,
Koreth-Funtoo and Koreth-Gentoo should actually be one and the same. So
I'm renaming this repo 'koreth'.